Docs: fix and standardize Maple and RET6 hardware pages.
Pick a standard layout for the information and fix mistakes.  Looks
like some of these errors were copy-pasted in from the old HTML
documentation, some crept in due to API changes, and some were just
wrong.

I've checked all the claims against the datasheets and source code, so
we should be in good shape now.